requirements.txt
streamlit
streamlit_chat

python-dotenv
chromadb

logging
warnings
operator
typing
ast

pandas
numpy

google-generativeai
langchain
langchain_community
langchain_core
langchain_google_genai

sentence-transformers==2.2.2

mysql-connector-python
pymysql